refactor: move auth mutation protocol owners
