+++ packages/shared-server/rallar-system/auth/mutation/auth-mutation-rejected-error.ts
@@ -1,0 +1,11 @@
+export class AuthMutationRejectedError extends Error {
+  readonly code = 'auth-mutation-rejected';
+
+  constructor(
+    message: string,
+    readonly status = 409,
+  ) {
+    super(message);
+    this.name = 'AuthMutationRejectedError';
+  }
+}
