+++ packages/shared-server/rallar-system/services/auth-state-mutations.ts
@@ -1,4 +1,5 @@
-export { decodeAuthMutationCommand, decodeAuthMutationResult } from './auth-state-codecs.ts';
+export { decodeAuthMutationCommand } from '../auth/mutation/decode-auth-mutation-command.ts';
+export { decodeAuthMutationResult } from '../auth/mutation/decode-auth-mutation-result.ts';
 export type {
     AuthComputedSession,
     AuthMutationCommand,
@@ -16,7 +17,7 @@ export type {
     IssueAuthWsTicketCommand,
     LogoutAuthSessionCommand,
     RegisterAuthUserCommand,
-} from './auth-state-contracts.ts';
-export { AuthMutationRejectedError } from './auth-state-errors.ts';
+} from '../auth/mutation/auth-mutation-contracts.ts';
+export { AuthMutationRejectedError } from '../auth/mutation/auth-mutation-rejected-error.ts';
 export { captureAuthMutationFacts } from './auth-state-read.ts';
 export { createAuthMutationService } from './auth-state-service.ts';
